pyproject.toml
[tool.poetry]
name = "demoshop"
version = "0.1.0"
description = ""
authors = ["Irina Kirillova <{ID}+{username}@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selene = "2.0.0rc4"
requests = "2.31.0"
allure-pytest = "2.13.2"
jsonschema = "4.20.0"
pytest = "7.4.3"
logging = "0.4.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
